refactor: split conjugations testing away from the main view
