+++ src/domains/conjugations/components/ConjugationsView.tsx
@@ -1,53 +1,34 @@
-import type { Component } from 'solid-js';
-import { Show } from 'solid-js';
+import { Show, type Component } from 'solid-js';
 import { createStore } from 'solid-js/store';
 import type { ConjugationLanguageCode } from '~/model/lang';
-import { ConjugationsResults } from './ConjugationsResults';
-import type { TensesValidations } from './ConjugationsTester';
-import { ConjugationsTester } from './ConjugationsTester';
 import { VerbInput } from './VerbInput';
-import { TenseFilter } from './tense-filter/TenseFilter';
-import type { Tense, VerbConjugations } from '../resources/conjugations-api';
+import type { VerbConjugations } from '../resources/conjugations-api';
 import { fetchVerbixConjugations } from '../resources/conjugations-api';
-
-interface ConjugationsViewState {
-  verbConjugations?: VerbConjugations;
-  selectedTenses: Tense[];
-  tensesResults: TensesValidations;
-  practiceIncorrect: boolean;
-  testingDone: boolean;
-  verbLoading: boolean;
-}
+import { ConjugationsTestView } from './ConjugationsTestView';
 
 const LANG_STORAGE_KEY = 'conjugations-lang-storage-key';
 
-const getInitialState: () => ConjugationsViewState = () => ({
-  verbConjugations: undefined,
-  selectedMoods: [],
-  selectedTenses: [],
-  tensesResults: {},
-  practiceIncorrect: false,
-  testingDone: false,
-  verbLoading: false,
-});
+type State = {
+  mode: 'search' | 'test';
+  verb: string;
+  verbLoading: boolean;
+  verbConjugations: VerbConjugations | undefined;
+};
 
 export const ConjugationsView: Component = () => {
   let verbInputEl: HTMLInputElement | undefined;
-
-  const [state, setState] =
-    createStore<ConjugationsViewState>(getInitialState());
-
-  const incorrectTenses = () =>
-    state.verbConjugations?.tenses.filter(t =>
-      state.tensesResults[t.name]?.some(v => v.valid === false)
-    );
-
-  const selectedTenses = () =>
-    state.practiceIncorrect ? incorrectTenses() : state.selectedTenses;
+  const [state, setState] = createStore<State>({
+    mode: 'search',
+    verb: '',
+    verbLoading: false,
+    verbConjugations: undefined as VerbConjugations | undefined,
+  });
 
   const applyVerb = async (verb: string, lang: ConjugationLanguageCode) => {
-    reset();
-    setState('verbLoading', true);
+    setState({
+      verb,
+      verbLoading: true,
+    });
 
     const verbConjugations = await fetchVerbixConjugations(lang, verb);
     setState({
@@ -56,41 +37,8 @@ export const ConjugationsView: Component = () => {
     });
   };
 
-  const reset = () => setState(getInitialState());
-
-  const selectTenses = (selectedTenses: Tense[]) => {
-    setState({ selectedTenses });
-
-    if (state.testingDone) {
-      setState('testingDone', false);
-    }
-  };
-
   const onLangChange = (lang: ConjugationLanguageCode) => {
     localStorage.setItem(LANG_STORAGE_KEY, lang);
-  };
-
-  const onTestingDone = (validationResults: TensesValidations) => {
-    setState({ testingDone: true, tensesResults: validationResults });
-  };
-
-  const onTryAgain = () => {
-    setState({
-      tensesResults: {},
-      testingDone: false,
-      practiceIncorrect: false,
-    });
-  };
-
-  const onTryDifferent = () => {
-    setState({
-      ...getInitialState(),
-      verbConjugations: state.verbConjugations,
-    });
-  };
-
-  const onPracticeIncorrect = () => {
-    setState({ testingDone: false, practiceIncorrect: true });
   };
 
   return (
@@ -114,38 +62,15 @@ export const ConjugationsView: Component = () => {
           </a>
         </span>
       </div>
-      <Show when={!state.testingDone}>
-        <div class="mt-4 grid items-start gap-4 md:grid-cols-[1fr_2fr]">
-          <Show when={!state.practiceIncorrect}>
-            <Show when={state.verbConjugations}>
-              {verbConjugations => (
-                <div class="rounded-lg bg-white p-4 shadow-md">
-                  <TenseFilter
-                    tenses={verbConjugations().tenses}
-                    selectedTenses={state.selectedTenses}
-                    onSelectedTenses={selectTenses}
-                  />
-                </div>
-              )}
-            </Show>
+      <Show when={state.verbConjugations}>
+        {conjugations => (
+          <Show when={state.mode === 'search'}>
+            <ConjugationsTestView
+              verb={state.verb}
+              verbConjugations={conjugations()}
+            />
           </Show>
-          <Show when={selectedTenses()?.length}>
-            <div class="rounded-lg bg-white p-4 shadow-md">
-              <ConjugationsTester
-                selectedTenses={selectedTenses()!}
-                onDone={onTestingDone}
-              />
-            </div>
-          </Show>
-        </div>
-      </Show>
-      <Show when={state.testingDone}>
-        <ConjugationsResults
-          tensesValidations={state.tensesResults}
-          onTryAgain={onTryAgain}
-          onTryDifferent={onTryDifferent}
-          onPracticeIncorrect={onPracticeIncorrect}
-        />
+        )}
       </Show>
     </div>
   );
